+++ CHANGES.rst
@@ -11,6 +11,7 @@ New Features:
 * index terms can be StyledText now (in addition to str)
 * the 'document author' metadata entry can now be displayed using a Field
 * Sphinx frontend: support the 'desc_signature_line' node (new in Sphinx 1.5)
+* rinoh --docs: open the online documentation in the default browser
 
 Changed: